Get current location - Added support for getting current device location to load restaurants nearby. - Also fixed tests which got out of date after adding the favorites feature.

+++ app/src/main/java/com/interview/doordashlite/datalayer/DataRepositoryImpl.kt
@@ -25,8 +25,8 @@ class DataRepositoryImpl: DataRepository, KoinComponent {
     private val sharedPreferences by inject<SharedPreferences>()
 
     override fun getRestaurantList(
-        latitude: Float,
-        longitude: Float
+        latitude: Double,
+        longitude: Double
     ) = doordashApi.getRestaurantList(latitude, longitude)
 
     override fun getRestaurant(id: String) = doordashApi.getRestaurant(id)
